feat(apiAdapter): universityAdapter base class corrected
