+++ apps/backend/src/ApiService/Adapter/NWU_Adapter/NWU_Adapter.ts
@@ -1,11 +1,8 @@
-import {
-  NotImplementedException,
-  RequestTimeoutException,
-} from '@nestjs/common';
+import { NotImplementedException } from '@nestjs/common';
 import { University_Adapter } from '../University_Adapter';
 import { CreateCourseDto } from 'src/Course/dto/course.dto';
-import { ModuleListResponseDto } from 'src/Module/dto/module.dto';
-import { EventListResponseDto } from 'src/Events/dto/EventDto.dto';
+import { CreateModuleDto } from 'src/Module/dto/module.dto';
+import { CreateEventDto } from 'src/Events/dto/EventDto.dto';
 
 interface ExternalCourse {
   course_id: string;
@@ -14,11 +11,10 @@ interface ExternalCourse {
   description: string;
 }
 
-export class NWU_Adapter implements University_Adapter {
-  constructor(
-    private readonly baseUrl: string,
-    private readonly uniId: string,
-  ) {}
+export class NWU_Adapter extends University_Adapter {
+  constructor(baseUrl: string, uniId: string) {
+    super(baseUrl, uniId);
+  }
 
   async authenticate(): Promise<void> {}
 
@@ -32,62 +28,14 @@ export class NWU_Adapter implements University_Adapter {
     // return courses;
 
     return result;
+  }
 
+  async getModules(): Promise<CreateModuleDto[]> {
     throw new NotImplementedException();
   }
 
-  async getModules(): Promise<ModuleListResponseDto> {
+  async getEvents(): Promise<CreateEventDto[]> {
     throw new NotImplementedException();
-  }
-
-  async getEvents(): Promise<EventListResponseDto> {
-    throw new NotImplementedException();
-  }
-
-  async request<T = any>(url: string): Promise<T> {
-    const timeout = 10000; //10 seconds
-    const controller = new AbortController();
-    const timeoutId = setTimeout(() => controller.abort(), timeout);
-
-    try {
-      let finalUrl = `${this.baseUrl}/${url}`;
-      finalUrl = finalUrl.endsWith('/') ? finalUrl : `${finalUrl}/`;
-
-      //example: url='courses'
-      const response = await fetch(`${finalUrl}`, {
-        signal: controller.signal,
-        headers: {
-          'Content-Type': 'application/json',
-          Accept: 'application/json',
-        },
-      });
-
-      clearTimeout(timeoutId);
-
-      //someting wrong
-      if (!response.ok) {
-        throw new Error(
-          `API request failed: ${response.status} | ${response.statusText}`,
-        );
-      }
-
-      const contentType = response.headers.get('content-type');
-      if (contentType?.includes('application/json')) {
-        return await response.json();
-      } else {
-        return (await response.text()) as T;
-      }
-    } catch (error) {
-      clearTimeout(timeoutId);
-
-      if (error instanceof Error && error.name === 'AbortError') {
-        throw new RequestTimeoutException(
-          `Request timed out after ${timeout}ms`,
-        );
-      }
-
-      throw error;
-    }
   }
 
   // 🎅's little helpers
