+++ apps/backend/src/ApiService/Registry/AdapterRegistry.ts
@@ -5,7 +5,6 @@ import {
 } from '@nestjs/common';
 import { University_Adapter } from '../Adapter/University_Adapter';
 import { UniversityDto } from '../../University/dto/university.dto';
-import { Example_Adapter } from '../Adapter/Example_Adapter/Example_Adapter';
 import { NWU_Adapter } from '../Adapter/NWU_Adapter/NWU_Adapter';
 
 @Injectable()
@@ -37,16 +36,17 @@ export class AdapterRegistry {
   } //END_register
 
   private createAdapter(uni: UniversityDto): University_Adapter {
-    if (uni.ApiIdentifier?.toUpperCase() === `UP`) {
-      const baseUrl = uni.BaseApiUrl ?? null;
-      const apiKey = uni.ApiKey ?? null;
+    // if (uni.ApiIdentifier?.toUpperCase() === `UP`) {
+    //   const baseUrl = uni.BaseApiUrl ?? null;
+    //   const apiKey = uni.ApiKey ?? null;
 
-      if (baseUrl === null || apiKey === null)
-        throw new BadRequestException(
-          `BaseUrl and ApiKey does not exist for uni[${JSON.stringify(uni)}]`,
-        );
-      else return new Example_Adapter(baseUrl, apiKey);
-    } else if (uni.ApiIdentifier?.toUpperCase() === 'NWU') {
+    //   if (baseUrl === null || apiKey === null)
+    //     throw new BadRequestException(
+    //       `BaseUrl and ApiKey does not exist for uni[${JSON.stringify(uni)}]`,
+    //     );
+    //   else return new Example_Adapter(baseUrl, apiKey);
+    // } else
+    if (uni.ApiIdentifier?.toUpperCase() === 'NWU') {
       const baseUrl = uni.BaseApiUrl ?? null;
 
       if (baseUrl === null)
